Clean Up Form Code; Fix Redirect Bug
There are a lot of little changes:

On form.erb the html, body, head, and title tags were added. Extra
scripts not used were deleted. Indenting was fixed. The form-group divs
were removed since they weren't doing anything.

For form.js, the code in script.js was moved there, and script.js was
deleted. Code was rearanged to look better. Lastly, a bug was fixed that
caused the page to redirect immediately after submiting the form. Now
when the form is submitted, the ajax waits for a response. Once the
response is received, the page gives a successful alert. Once the user
presses OK, they are then redirected to the main page.

form.html was deleted since form.erb is being used instead.

index.html had it's link to script.js removed.